refactor(desktop): give the durable write its own module
writeDurableFile is a generic filesystem primitive, not checkpoint
domain logic, and profile.ts became its second consumer when the
workspace writes went atomic. Code review flagged the ownership: the
synchronous durable write now lives in src/durable-write.ts and both
consumers import it from there, while the checkpoint writer keeps its
directory-mode conventions in place. No behavior change.
